Code cleanup. Start implementing gHTC.
Some reformatting.
More precise variable types.

+++ firmware/humidity.h
@@ -22,4 +22,4 @@
 
 void adc_setup (void);
 void adc_startc (short pin);
-int mhumid (short pin);
+uint16_t mhumid (short pin);
